wait_for_input, write_to_file -> agent property instead of module property
